Split tests based on DPE commands
